Refactor to make spark a first class citizen of com.linkedin.venice
