+++ clients/venice-push-job/src/test/java/com/linkedin/venice/hadoop/spark/input/hdfs/TestSparkInputFromHdfs.java
@@ -11,8 +11,8 @@ import static com.linkedin.venice.hadoop.VenicePushJobConstants.SCHEMA_STRING_PR
 import static com.linkedin.venice.hadoop.VenicePushJobConstants.UPDATE_SCHEMA_STRING_PROP;
 import static com.linkedin.venice.hadoop.VenicePushJobConstants.VALUE_FIELD_PROP;
 import static com.linkedin.venice.hadoop.VenicePushJobConstants.VSON_PUSH;
-import static com.linkedin.venice.hadoop.spark.SparkConstants.DEFAULT_SCHEMA;
-import static com.linkedin.venice.hadoop.spark.input.hdfs.VeniceHdfsInputTable.INPUT_TABLE_NAME;
+import static com.linkedin.venice.spark.SparkConstants.DEFAULT_SCHEMA;
+import static com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputTable.INPUT_TABLE_NAME;
 import static com.linkedin.venice.utils.TestWriteUtils.DEFAULT_USER_DATA_VALUE_PREFIX;
 import static com.linkedin.venice.utils.TestWriteUtils.NAME_RECORD_V2_SCHEMA;
 import static com.linkedin.venice.utils.TestWriteUtils.STRING_TO_STRING_SCHEMA;
@@ -23,6 +23,13 @@ import com.linkedin.venice.schema.vson.VsonAvroSerializer;
 import com.linkedin.venice.schema.writecompute.WriteComputeSchemaConverter;
 import com.linkedin.venice.serializer.FastSerializerDeserializerFactory;
 import com.linkedin.venice.serializer.RecordDeserializer;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputPartition;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputPartitionReader;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputPartitionReaderFactory;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputScan;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputScanBuilder;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsInputTable;
+import com.linkedin.venice.spark.input.hdfs.VeniceHdfsSource;
 import com.linkedin.venice.utils.TestWriteUtils;
 import java.io.File;
 import java.io.IOException;
